services/energy/requirements.txt
# Energy Service Dependencies

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uuid7==0.1.0
alembic>=1.12.0
prometheus-client==0.19.0
slowapi==0.1.9
pyjwt==2.8.0
